Revert "Revert "Use standard JUnit structure in the SysAtt110 test""
This reverts commit 884923b658e0252ae40b12fb1790c75eb7762f04.
